TST: Add Missing Import
